Move protocols to separate module
Also only import what is needed at runtime.
